+++ src/main/java/main/Application.java
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 @SpringBootApplication
-@EnableMongoRepositories("repositories")
-@ComponentScan({"repositories", "controllers", "domain", "services"})
+@EnableMongoRepositories("dbmanagement")
+@ComponentScan({"repositories", "parser", "dbmanagement", "agents", "domain", "services"})
 public class Application {
 
     public static void main(String[] args) {
