+++ src/test/java/persistance_tests/DatabaseTest.java
@@ -9,10 +9,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
+import dbmanagement.AgentsRepository;
+import dbmanagement.Database;
 import domain.Agent;
 import main.Application;
-import repositories.AgentsRepository;
-import repositories.Database;
 import util.JasyptEncryptor;
 
 /**
